Inline shareIntent hook
