+++ src/app/_layout.tsx
@@ -5,8 +5,9 @@ import {
   useFonts,
 } from "@expo-google-fonts/source-sans-3";
 import { StripeProvider } from "@stripe/stripe-react-native";
-import { Stack, useSegments } from "expo-router";
-import { ShareIntentProvider } from "expo-share-intent";
+import * as Linking from "expo-linking";
+import { Href, Stack, useRouter, useSegments } from "expo-router";
+import { ShareIntentProvider, useShareIntentContext } from "expo-share-intent";
 import * as SplashScreen from "expo-splash-screen";
 import { useEffect } from "react";
 import { LogBox, Platform } from "react-native";
@@ -22,12 +23,12 @@ import View from "#/components/design/View";
 import MissionPopup from "#/components/popups/MissionPopup";
 import ToastShareSheet from "#/components/popups/ToastShareSheet";
 import Colors from "#/constants/Colors";
+import Config from "#/constants/Config";
 import NotificationManager from "#/helpers/Notifications";
 import PersonalStore from "#/helpers/Stores/PersonalStore";
 import { BadgeProvider } from "#/helpers/provider/BadgeProvider";
 import { SettingsProvider } from "#/helpers/provider/SettingsProvider";
 import useAppColorScheme from "#/hooks/useAppColorScheme";
-import { useHandleShareIntent } from "#/hooks/useHandleShareIntent";
 import { useNotificationObserver } from "#/hooks/useNotificationObserver";
 
 // Hide warning for new native event emitter
@@ -166,10 +167,45 @@ export const unstable_settings = {
 };
 
 /**
- * Tiny local component that calls the hook and renders nothing.
+ * Inline runner that handles incoming share intents and routes them.
  */
 const ShareIntentRunner = () => {
-  useHandleShareIntent();
+  const router = useRouter();
+  const { shareIntent, hasShareIntent } = useShareIntentContext();
+
+  useEffect(() => {
+    if (!hasShareIntent || !shareIntent) return;
+
+    // Delay routing to allow root layout to finish mounting
+    const t = setTimeout(() => {
+      if (shareIntent?.type === "weburl") {
+        try {
+          const { path } = Linking.parse(shareIntent.webUrl);
+          if (!shareIntent.webUrl.includes(Config.wpUrl)) {
+            router.push({
+              pathname: "/search",
+              params: { tag: shareIntent.webUrl },
+            });
+            return;
+          }
+
+          const safePath =
+            typeof path === "string" && path.length > 0
+              ? path.startsWith("/")
+                ? path
+                : `/${path}`
+              : "/search";
+
+          router.push(safePath as Href);
+        } catch {
+          // swallow malformed urls
+        }
+      }
+    }, 0);
+
+    return () => clearTimeout(t);
+  }, [hasShareIntent, router, shareIntent]);
+
   return null;
 };
